requirements.txt
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fonttools==4.48.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pillow==10.2.0
PyGLM==2.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
setuptools==69.0.3
shapely==2.0.3
six==1.16.0
